+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 ## Unreleased
+- Fix possibly unhandled `Runtime Exception` on `Advance` or `Finish` during not-started `Progress`
 
 ## 1.0.0 - YYYY-DD-MM
 ### Added